fixed bug; Do not creating a file with a name used
Now, when you try to create a file with a name that
already exists, the konsole will ask if you want to
edit the old file or exit the program to be able to
create a new file with a different name.

The old version of this script, erased the old file
and created a new empty file, losing all the past
work.